java.lang.UnsupportedOperationException example
All Java Collections extend Collection interface. So, they have to implement all the methods in the Collection interface. However, certain Java collections are optimized to be used in specific conditions and do not support all the Collection operations (methods). When an unsupported operation is calledon a Collection, the Collection Implementation would throw anUnsupportedOperationException